PUT(ID yang belum ada tidak bisa dimasukkan)

+++ src/main/java/ws/a/restful/ProductServiceController.java
@@ -5,6 +5,7 @@
  */
 package ws.a.restful;
 
+import exception.ProductNotfoundException;
 import java.util.HashMap;
 import java.util.Map;
 import model.Product;
@@ -42,6 +43,7 @@ public class ProductServiceController {
     
     @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
     public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) {
+        if(!productRepo.containsKey(id))throw new ProductNotfoundException();
         productRepo.remove(id);
         product.setId(id);
         productRepo.put(id, product);
